delete route functional

+++ Controllers/LibraryController.cs
@@ -3,6 +3,7 @@ using Librestack.Models;
 using Librestack.Services;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Superpower.Model;
 
 namespace Librestack.Controllers;
 
@@ -21,14 +22,33 @@ public class LibraryController : ControllerBase
     [Authorize]
     public async Task<ActionResult<List<Library>>> GetLibrary()
     {
-        return await _interfaceLibraryService.GetLibrary();
+        var result = await _interfaceLibraryService.GetLibrary();
+        if (result is null)
+            return BadRequest("Library does not exist");
+
+        if (result.Count() == 0)
+            return BadRequest("Library has no entries");
+
+        if (result.Count() > 1)
+            return StatusCode(418);
+
+        return result;
     }
 
     [HttpGet("getLibraryEntry")]
     [Authorize]
     public async Task<ActionResult<Library>> GetLibraryEntry(int id)
     {
-        return await _interfaceLibraryService.GetLibraryEntry(id);
+        Console.WriteLine($"getLibraryEntry - int value: {id}");
+
+        if (id == 0)
+            return BadRequest("id parameter cannot be 0");
+
+        var result = await _interfaceLibraryService.GetLibraryEntry(id);
+        if (result is null)
+            return BadRequest($"id {id} not found");
+
+        return result;
     }
 
     [HttpPost("updateLibraryEntry")]
@@ -39,7 +59,7 @@ public class LibraryController : ControllerBase
 
         var updated = await _interfaceLibraryService.UpdateLibraryMetaData(library);
         if (!updated)
-            return NotFound();
+            return NotFound($"title {library.Title} not found");
 
         var result = new { message = "Library entry updated" };
         return new JsonResult(result);
@@ -60,12 +80,16 @@ public class LibraryController : ControllerBase
     [Authorize]
     public async Task<IActionResult> DeleteLibraryEntry(int id)
     {
+        if (id == 0)
+            return BadRequest("id parameter cannot be 0");
+
+        var data = await _interfaceLibraryService.GetLibraryEntry(id);
         var deleted = await _interfaceLibraryService.DeleteLibraryEntry(id);
 
-        if (!deleted)
-            return NotFound();
+        if (!deleted || data is null)
+            return NotFound($"id {id} not found in database");
 
-        return NoContent();
+        return Ok($"{data.Title} removed from database");
     }
 
 }
